resloved screen navigation
